+++ __tests__/controllers/modules/object.test.js
@@ -6,7 +6,7 @@ const { checkNativeBrand, cloneDeep, mergeDeep, assignDeep } = require('../../..
 // test
 describe('Check the ENV', () => {
     test('Should run in test mode', () => {
-        expect(process.env.NODE_ENV).toEqual('test');
+        expect(process.env['NODE_ENV']).toEqual('test');
     });
 });