+++ __tests__/controllers/search.test.js
@@ -7,7 +7,7 @@ const { getAggregationQuery, getFilterExp,
 // test
 describe('Check the ENV', () => {
     test('Should run in test mode', () => {
-        expect(process.env.NODE_ENV).toEqual('test');
+        expect(process.env['NODE_ENV']).toEqual('test');
     });
 });