+++ models/configs.js
@@ -6,15 +6,16 @@ const
 // ==============================
 //  SCHEMA
 // ==============================
-const settingModelSchema    = new mongoose.Schema({
+const configModelSchema    = new mongoose.Schema({
     active                  : { type: Boolean, default: false, unique: true },
     title                   : { type: String, default: 'New Website' },
     description             : { type: String, default: 'n/a' },
     keywords                : { type: String, default: 'n/a' },
     sets: {
         imageTypes          : { type: [[String]], default: ['gif', 'jpe?g', 'png', 'svg', 'tiff', 'webp']},
-        timeFormat          : { type: String, default: '' },
+        language            : { type: String, default: 'en' },
         timezone            : { type: String, default: '' },
+        timeFormat          : { type: String, default: '' },
         sort                : { type: String, default: '' },
         num                 : { type: Number, default: 5 },
     },
@@ -29,17 +30,21 @@ const settingModelSchema    = new mongoose.Schema({
 //  STATIC METHODS
 // ==============================
 // initialization (model)
-settingModelSchema.static('initialize', async function(next = () => {}) {
-    return await this.findOne({ active: true }) ? next() : await this.create({ active: true }).then(() => next());
+configModelSchema.static('initialize', async function(next = () => {}) {
+    process.env['$WEBSITE_CONFIGS'] = JSON.stringify(await this.findOne({ active: true }));
+    if (process.env['$WEBSITE_CONFIGS'] === 'null') {
+        process.env['$WEBSITE_CONFIGS'] = JSON.stringify(await this.create({ active: true }));
+    } return next();
 });
 
 
 // update settings (model)
-settingModelSchema.static('updateSettings', function(dataToBeUpdated, next) {
-    return this.findOneAndUpdate({ active: true }, dataToBeUpdated, { new: true }, next);
+configModelSchema.static('updateSettings', async function(doc, next = () => {}) {
+    process.env['$WEBSITE_CONFIGS'] = JSON.stringify(await this.findOneAndUpdate({ active: true }, doc, { new: true }));
+    return next();
 });
 
 
 
 // exports
-module.exports = mongoose.model('settings', settingModelSchema);
+module.exports = mongoose.model('settings', configModelSchema);