+++ routers/index.js
@@ -11,7 +11,7 @@ const
 function _useMiddleware(app) {
     // external
     //app.use(favicon(path.join(__dirname, 'public', 'favicon.ico')));
-    if (process.env.NODE_ENV === 'dev') app.use(logger('dev'));
+    if (process.env['NODE_ENV'] === 'dev') app.use(logger('dev'));
     app.use(methodOverride('_method'));
     app.use(bodyParser.urlencoded({ extended: true }));
     app.use(bodyParser.json());
@@ -24,7 +24,7 @@ function _useMiddleware(app) {
 
 function _useRoutes(app) {
     // seed
-    if (process.env.NODE_ENV === 'dev' || 'test') app.use('/seed', require('./seed'));
+    if (process.env['NODE_ENV'] === 'dev' || 'test') app.use('/seed', require('./seed'));
 
     // units
     app.use('/home', require('./admin'));